pyproject.toml
[project]
name = "resume-ner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "huggingface-hub>=1.0.1",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "mypy>=1.18.2",
    "ruff>=0.14.3",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.7",
    "nbformat>=5.10.4",
    "plotly>=6.3.1",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.4",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "spacy>=3.8.11",
    "scikit-learn>=1.7.2",
    "pillow>=10.1.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]
